// src/types.ts
// Definición de los tipos de pestañas disponibles
export type TabType = 'overview' | 'gdp' | 'researchers' | 'patents';

// Claves de traducción disponibles
export type TranslationKey = 
  | 'overview' 
  | 'gdp' 
  | 'researchers' 
  | 'patents'
  | 'app.title'
  | 'app.subtitle'
  | 'app.developedFor'
  | 'app.developedBy'
  | 'app.lastUpdate'
  | 'nav.overview'
  | 'nav.gdp'
  | 'nav.researchers'
  | 'nav.patents'; 